sureEsma.py
import requests
from bs4 import BeautifulSoup
import arabic_reshaper
from bidi.algorithm import get_display


sureler = {
    11: "HACC",
    14: "TAHA",
    15: "HUD",
    26: "HADID",
    36: "BELED",
    51: "MAIDE",
    54: "NEBE",
    57: "HUMEZE",
    63: "SEBE",
    64: "NUH",
    65: "ENBIYA",
    70: "YASIN, LEYL",
    72: "SECDE",
    75: "ZILZAL",
    77: "BEYYINE",
    83: "MUCADDELE",
    88: "NAHL",
    90: "MULK, SAD",
    93: "NECM",
    100: "KAF",
    103: "CIN",
    105: "KEHF",
    111: "ALA, NAS",
    112: "NISA",
    118: "CUMA",
    120: "NEML, FIL",
    124: "MUZZEMMIL",
    126: "YUNUS",
    132: "MUHAMMED, ABESE",
    137: "MU-MIN",
    140: "TALAK",
    155: "KIYAMET",
    156: "YUSUF",
    162: "ENFAL, INSAN, ENAM",
    167: "MAUN",
    170: "KALEM",
    182: "VAKIA",
    190: "AHKAF",
    193: "MUMINUN",
    200: "ALAK",
    210: "FELAK",
    211: "HICR, BURUC",
    214: "HAKKA",
    215: "TUR",
    221: "LOKMAN",
    246: "RUM",
    247: "ZUMER",
    250: "SAFF",
    256: "NUR",
    258: "IBRAHIM",
    263: "ISRA",
    274: "RAD",
    280: "KASAS",
    283: "FECR",
    290: "MERYEM, FATIR",
    298: "RAHMAN",
    304: "KADIR",
    307: "BAKARA",
    310: "TARIK",
    314: "MEARIC",
    327: "MUNAFIKUN",
    340: "KAMER, NASR",
    341: "INFITAR",
    349: "MUTAFFIFIN",
    352: "A RAF",
    357: "KAFIRUN",
    360: "ASR",
    376: "KARIA",
    391: "ALI IMRAN",
    400: "ŞEMS",
    413: "TEVBE",
    431: "FURKAN",
    460: "TIN",
    486: "ADIYAT",
    488: "FETIH",
    494: "FATIHA",
    508: "HAŞR",
    516: "ŞURA",
    519: "CASIYE",
    529: "NAZIAT",
    543: "MUMTEHINE",
    548: "ANKEBUT",
    552: "INŞIKAK",
    559: "INŞIRAH",
    572: "ŞUARA",
    610: "KUREYŞ",
    612: "HUCURAT",
    636: "TEKVIR",
    652: "SAFFAT",
    655: "DUHAN",
    658: "TAHRIM",
    690: "FUSSILET",
    712: "AHZAB",
    722: "IHLAS",
    726: "KEVSER",
    731: "MURSELAT",
    804: "TEBBET",
    818: "DUHA",
    893: "ZUHRUF",
    1121: "TEKASUR",
    1248: "MUDDESSIR",
    1312: "ZARIYAT",
    1316: "GAŞIYE",
    1453: "TEGABUN"
}

esmalar = {
    13: "Ehad",
    14: "Vehhab, Vacid",
    18: "Hayy",
    19: "Vahid",
    20: "Vedud, Hâdi",
    37: "Evvel",
    46: "Velî",
    47: "Vâli",
    48: "Mâcid",
    55: "Mucîb",
    56: "Mübdî",
    57: "Mecîd",
    62: "Bâtın, Hâmid",
    66: "Vekîl, Allah",
    68: "Muhyî, Hakem",
    72: "Bâsit",
    73: "Celîl",
    78: "Hakîm",
    80: "Hasîb",
    86: "Bedî",
    88: "Halîm",
    90: "Melik",
    94: "Aziz",
    104: "Adl",
    108: "Hakk",
    110: "Aliyy",
    113: "Baki",
    114: "Cami",
    116: "Kaviyy",
    117: "Muizz",
    124: "Muid",
    129: "Latîf",
    131: "Selam",
    134: "Samed",
    137: "Mü-min, Vâsi",
    145: "Müheymin",
    148: "Muhsî",
    150: "Alîm",
    156: "Kayyûm, Afuvv",
    161: "Manî",
    170: "Kuddûs",
    180: "Semî",
    184: "Mukaddim",
    201: "Nâfî",
    202: "Berr",
    206: "Cebbar",
    209: "Muksit",
    212: "Malik-ul Mülk",
    214: "Bâri",
    232: "Kebîr",
    256: "Nur",
    258: "Rahîm",
    270: "Kerîm",
    287: "Raûf",
    298: "Rahman, Sabûr",
    302: "Basîr",
    305: "Kâdir",
    306: "Kahhâr",
    308: "Rezzâk",
    312: "Rakîb",
    319: "Şehîd",
    336: "Musavvir",
    351: "Râfi",
    409: "Tevvâb",
    489: "Fettah",
    490: "Mümît",
    500: "Metîn",
    514: "Reşîd",
    526: "Şekûr",
    550: "Mükît",
    551: "Müteali",
    573: "Bâis",
    630: "Muntakim",
    662: "Mütekebbir",
    707: "Vâris",
    731: "Hâlik",
    744: "Muktedir",
    770: "Muzill",
    801: "Ahir",
    812: "Habîr",
    847: "Muahhir",
    903: "Kâbid",
    998: "Hafîz",
    1001: "Dârr",
    1020: "Azîm",
    1060: "Ganiyy",
    1100: "Zülcelâli vel-İkrâm, Muğnî",
    1106: "Zâhir",
    1286: "Gafûr",
    1481: "Hâfız",
}


def get_ebced_and_arabic(url):
    # Web sayfasına istek gönderme
    response = requests.get(url)
    
    # İstek başarılı mı kontrol etme
    if response.status_code == 200:
        # HTML içeriğini alma
        html_content = response.text
        soup = BeautifulSoup(html_content, 'html.parser')
        
        # Tüm 'Sonebced' id'li div etiketlerini bulma
        sonebced_divs = soup.find_all('div', {'id': 'Sonebced'})
        
        ebced_value = None
        arabic_text = None
        
        for div in sonebced_divs:
            h5_tag = div.find('h5')
            if h5_tag:
                h5_text = h5_tag.text.strip()
                span_text = div.find('span').text.strip()
                
                if h5_text == 'Ebced Değeri':
                    ebced_value = span_text
                elif h5_text == 'Arapça Yazılışı':
                    arabic_text = span_text # Arapça metni ters çevirme
                    reshaped_text = arabic_reshaper.reshape(arabic_text)
                    arabic_text = get_display(reshaped_text)
        
        return ebced_value, arabic_text
    else:
        return None, None

# URL

"""isim = input("İsminizi girin: ")
anne_ismi = input("Annenizin ismini girin: ")
baba_ismi = input("Babanızın ismini girin: ")

url = "https://www.ozgulyildiz.com/araclar/ebced/"

# Fonksiyonu çağırma ve ebced değeri ile Arapça yazılışı alma
ebced_degeri, arabic_text = get_ebced_and_arabic(url + isim)
print(f"Ebced Değeri: {ebced_degeri}")
print(f"Arapça Yazılışı: {arabic_text}")
"""


def dogumGunuToplama(birthdate):
    # Doğum tarihi 'gg aa yyyy' formatında olmalıdır.
    total_sum = 0
    for char in birthdate:
        if char.isdigit():
            total_sum += int(char)
    
    return total_sum

def en_yakin_esma_bul(number1, number2, number3):
    number = int(number1) + int(number2) + int(number3)

    keys = list(esmalar.keys())
    closest_key = min(keys, key=lambda x: abs(x - number))

    # Check if there is another close key
    closest_keys = [closest_key]
    keys.remove(closest_key)
    if keys:
        second_closest_key = min(keys, key=lambda x: abs(x - number))
        if abs(second_closest_key - number) == abs(closest_key - number):
            closest_keys.append(second_closest_key)

    return {key: esmalar[key] for key in closest_keys}

def en_yakin_sure_bul(number1, number2, number3):
    number = int(number1) + int(number2) + int(number3)

    keys = list(sureler.keys())
    closest_key = min(keys, key=lambda x: abs(x - number))

    # Check if there is another close key
    closest_keys = [closest_key]
    keys.remove(closest_key)
    if keys:
        second_closest_key = min(keys, key=lambda x: abs(x - number))
        if abs(second_closest_key - number) == abs(closest_key - number):
            closest_keys.append(second_closest_key)

    return {key: sureler[key] for key in closest_keys}


def akil_fikir_sayisi_hesaplama(number1, number2):
    try:
        number1 = int(number1)
        number2 = int(number2)
    except ValueError:
        return "Inputların ikisi de sayıya dönüştürülebilmelidir."

    toplam = number1 + number2
    kalan = toplam % 9
    if kalan == 0:
        kalan = 9
        
    return kalan

def calculate_arabic_ebced(name):
    """
    İsmin Arapça yazılışını ve ebced değerini hesaplar
    """
    # Bu fonksiyon örnek olarak basit bir dönüşüm yapıyor
    # Gerçek uygulamada daha kapsamlı bir Arapça dönüşüm sistemi kullanılmalı
    arabic_values = {
        'ahmet': ('احمد', 53),
        'nazife': ('نظيفة', 1045),
        'oğuz': ('زوغوا', 1020)
    }
    
    name = name.lower()
    if name in arabic_values:
        return arabic_values[name]
    return None, None

def calculate_fikir_sayisi(dogum_gunu):
    """
    Doğum gününden fikir sayısını hesaplar
    """
    # Örnek hesaplama - gerçek hesaplama algoritması uygulanmalı
    return "2"

def calculate_akil_zeka_sayisi(dogum_gunu):
    """
    Doğum gününden akıl/zeka sayısını hesaplar
    """
    # Örnek hesaplama - gerçek hesaplama algoritması uygulanmalı
    return "9"

def get_esma(dogum_gunu, isim):
    """
    Kişinin Esmasını hesaplar
    """
    # Örnek hesaplama - gerçek hesaplama algoritması uygulanmalı
    return "{1106: 'Zâhir'}"

def get_sure(dogum_gunu, isim):
    """
    Kişinin Suresini hesaplar
    """
    # Örnek hesaplama - gerçek hesaplama algoritması uygulanmalı
    return "{1121: 'TEKASUR'}"

def calculate_sure_esma(dogum_gunu, isim, anne_ismi, baba_ismi):
    """
    Tüm Sure ve Esma hesaplamalarını yapar
    """
    # İsmin Arapça yazılışı ve ebced değeri
    arabic_text, ebced_value = calculate_arabic_ebced(isim)
    
    results = {
        'isim': isim,
        'arabic_text': arabic_text,
        'ebced_value': ebced_value,
        'fikir_sayisi': calculate_fikir_sayisi(dogum_gunu),
        'akil_zeka_sayisi': calculate_akil_zeka_sayisi(dogum_gunu),
        'esma': get_esma(dogum_gunu, isim),
        'sure': get_sure(dogum_gunu, isim)
    }
    
    return results
