requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.31.0
beautifulsoup4==4.12.3
arabic-reshaper==3.0.0
python-bidi==0.4.2
gunicorn==20.1.0
Flask-Login==0.6.2
python-dotenv==1.0.0 